pyproject.toml
[project]
name = "chunker"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "gunicorn>=23.0.0",
    "langchain-text-splitters>=0.3.8",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "uvicorn[standard]>=0.34.2",
]

[tool.pytest.ini_options]
pythonpath = [
  "app"
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
]

[project.optional-dependencies]
cpu = [
    "sentence-transformers>=4.1.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]
gpu = [
    "sentence-transformers>=4.1.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "gpu" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "gpu" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "gpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
